Start working on size_t warnings on 32-bit

+++ CesiumVectorData/src/VectorRasterizer.cpp
@@ -90,7 +90,7 @@ VectorRasterizer::VectorRasterizer(
       imageHeight,
       BL_FORMAT_PRGB32,
       reinterpret_cast<void*>(pData),
-      (int64_t)imageWidth * (int64_t)this->_imageAsset->channels);
+      (size_t)imageWidth * (size_t)this->_imageAsset->channels);
 
   this->_context.begin(this->_image);
   // Initialize the image as all transparent.
